refactoring

+++ Mud_js/src/eventhandlers.js
@@ -7,15 +7,15 @@ function nameHandler(domElement, domWorkSpace) {
     event.preventDefault();
     clearSpace(domWorkSpace)
     clearSpace(domElement)
-    let string
+    let htmlString
     const id = parseInt(event.target.dataset.id)
     if (event.target.id === 'drink-name') {
-      string = Drink.getDrinkById(id).renderAll()
+      htmlString = Drink.getDrinkById(id).renderAll()
     } else if (event.target.id === 'order-name') {
-      string = Order.getOrderById(id).renderAll()
+      htmlString = Order.getOrderById(id).renderAll()
     }
-    if (string) {
-      renderSpace(domElement, string)
+    if (htmlString) {
+      renderSpace(domElement, htmlString)
     }
   }
 }
@@ -30,30 +30,35 @@ function clearSpace(domElement) {
   }
 }
 
-// refactor this below
 function showSpaceHandler(domWorkSpace, domShowSpace, domNameSpace ) {
   return event => {
     event.preventDefault();
     const id = parseInt(event.target.dataset.id)
-    switch(event.target.id) {
-      case 'edit-drink':
-        clearSpace(domWorkSpace)
-        renderSpace(domWorkSpace, Drink.getDrinkById(id).renderForm())
-        break;
-      case 'delete-drink':
-        clearSpace(domWorkSpace)
-        deleteDrink(id, domShowSpace)
-        break;
-      case 'edit-order':
-        clearSpace(domWorkSpace)
-        renderSpace(domWorkSpace, Order.getOrderById(id).renderForm())
-        Order.getOrderById(id).setFormDrinkQuantityValue()
-        break;
-      case 'delete-order':
-        clearSpace(domWorkSpace)
-        deleteOrder(id, domShowSpace)
-        break;
+    if (event.target.tagName === "BUTTON") {
+      clearSpace(domWorkSpace)
     }
+    if (event.target.parentElement.id === "drink-show") {
+      drinkShowSpaceHandler(domWorkSpace, domShowSpace, domNameSpace, id)
+    } else if (event.target.parentElement.id === "order-show") {
+      orderShowSpaceHandler(domWorkSpace, domShowSpace, domNameSpace, id)
+    }
+  }
+}
+
+function orderShowSpaceHandler(domWorkSpace, domShowSpace, domNameSpace, id) {
+  if (event.target.id === "edit-order") {
+    renderSpace(domWorkSpace, Order.getOrderById(id).renderForm());
+    Order.getOrderById(id).setFormDrinkQuantityValue()
+  } else if (event.target.id === 'delete-order') {
+    deleteOrder(id, domShowSpace)
+  }
+}
+
+function drinkShowSpaceHandler(domWorkSpace, domShowSpace, domNamesSpace, id) {
+  if (event.target.id === "edit-drink") {
+    renderSpace(domWorkSpace, Drink.getDrinkById(id).renderForm())
+  } else if (event.target.id === 'delete-drink') {
+    deleteDrink(id, domShowSpace)
   }
 }
 
@@ -118,45 +123,80 @@ function orderFormHandler(domShow, domOrderClients) {
 
 
 //refactor all of this for sure
+// function drinkOrderCheck() {
+//   const arr = [...document.getElementsByClassName('order-drink')]
+//   const orderId = parseInt(document.getElementById('edit-order').dataset.id)
+//   arr.forEach(element => {
+//       if (Order.getOrderById(orderId).drink_ids.includes(parseInt(element.dataset.id))) {
+//         const drinkOrder = DrinkOrder.getDoByOrderIdDrinkId(orderId, parseInt(element.dataset.id))
+//         if (parseInt(element.value) > 0 && drinkOrder.quantity !== parseInt(element.value)) {
+//          drinkOrder.quantity = parseInt(element.value)
+//          DrinkOrderAdapter.updateDrinkOrder(drinkOrder)
+//        } else if (parseInt(element.value) === 0 ) {
+//          DrinkOrderAdapter.deleteDrinkOrder(drinkOrder)
+//          Order.getOrderById(orderId).drink_orders = Order.getOrderById(orderId).drink_orders.filter(dO => dO.id !== drinkOrder.id )
+//          Order.getOrderById(orderId).drink_ids = Order.getOrderById(orderId).drink_ids.filter(id => id !== drinkOrder.drink_id)
+//        }
+//       } else if (parseInt(element.value) > 0) {
+//         const newDrinkOrder = new DrinkOrder({order_id: orderId, drink_id: parseInt(element.dataset.id), quantity: parseInt(element.value)})
+//         DrinkOrderAdapter.postNewDrinkOrder(newDrinkOrder).then(dO => {
+//           console.log(dO)
+//           setDataId(dO)
+//           setDrinkOrderIdDom(orderId, dO)
+//         })
+//         Order.getOrderById(orderId).drink_orders.push(newDrinkOrder);
+//         Order.getOrderById(orderId).drink_ids.push(parseInt(element.dataset.id))
+//
+//       }
+//     })
+// }
 function drinkOrderCheck() {
   const arr = [...document.getElementsByClassName('order-drink')]
   const orderId = parseInt(document.getElementById('edit-order').dataset.id)
+  const workOrder = Order.getOrderById(orderId)
   arr.forEach(element => {
-      if (Order.getOrderById(orderId).drink_ids.includes(parseInt(element.dataset.id))) {
-        const drinkOrder = DrinkOrder.getDoByOrderIdDrinkId(orderId, parseInt(element.dataset.id))
-        if (parseInt(element.value) > 0 && drinkOrder.quantity !== parseInt(element.value)) {
-         drinkOrder.quantity = parseInt(element.value)
+      const drinkId = parseInt(element.dataset.id)
+      const drinkQuantity = parseInt(element.value)
+      if (workOrder.drink_ids.includes(drinkId)) {
+        const drinkOrder = DrinkOrder.getDoByOrderIdDrinkId(orderId, drinkId)
+        if (drinkQuantity > 0 && drinkOrder.quantity !== drinkQuantity) {
+         drinkOrder.quantity = drinkQuantity
          DrinkOrderAdapter.updateDrinkOrder(drinkOrder)
-       } else if (parseInt(element.value) === 0 ) {
-         console.log(drinkOrder)
-         DrinkOrderAdapter.deleteDrinkOrder(drinkOrder)
-
-         Order.getOrderById(orderId).drink_orders = Order.getOrderById(orderId).drink_orders.filter(dO => dO.id !== drinkOrder.id )
-         Order.getOrderById(orderId).drink_ids = Order.getOrderById(orderId).drink_ids.filter(id => id !== drinkOrder.drink_id)
+       } else if (drinkQuantity === 0 ) {
+         deleteDrinkOrderFront(drinkOrder, workOrder)
+        //  DrinkOrderAdapter.deleteDrinkOrder(drinkOrder)
+        //  workOrder.drink_orders = workOrder.drink_orders.filter(dO => dO.id !== drinkOrder.id )
+        //  workOrder.drink_ids = workOrder.drink_ids.filter(id => id !== drinkOrder.drink_id)
        }
-      } else if (parseInt(element.value) > 0) {
-        const newDrinkOrder = new DrinkOrder({order_id: orderId, drink_id: parseInt(element.dataset.id), quantity: parseInt(element.value)})
-        DrinkOrderAdapter.postNewDrinkOrder(newDrinkOrder).then(dO => {
-          console.log(dO)
-          setDataId(dO)
-          setDrinkOrderIdDom(orderId, dO)
-        })
-        Order.getOrderById(orderId).drink_orders.push(newDrinkOrder);
-        Order.getOrderById(orderId).drink_ids.push(parseInt(element.dataset.id))
-
+     } else if (drinkQuantity > 0) {
+        newDrinkOrder(orderId, drinkId, drinkQuantity, workOrder)
       }
     })
 }
 
-function setDrinkOrderIdDom(order_id, drinkOrderObj) {
-  Order.getOrderById(order_id).drink_orders.map(dO => {
-    if (dO.id === undefined) {
-      console.log(dO)
-      console.log(drinkOrderObj)
-      dO.id = drinkOrderObj.id
-    }
-  })
+function deleteDrinkOrderFront(drinkOrderObj, orderObj) {
+  DrinkOrderAdapter.deleteDrinkOrder(drinkOrderObj).then(drinkOrderObj => DrinkOrder.deleteDrinkOrderMemory(drinkOrderObj))
+  orderObj.drink_orders = orderObj.drink_orders.filter(dO => dO.id !== drinkOrderObj.id )
+  orderObj.drink_ids = orderObj.drink_ids.filter(id => id !== drinkOrderObj.drink_id)
 }
+
+// function newDrinkOrder(orderId, drinkId, drinkQuantity, orderObj) {
+//   const newDrinkOrder = new DrinkOrder({order_id: orderId, drink_id: drinkId, quantity: drinkQuantity})
+//   DrinkOrderAdapter.postNewDrinkOrder(newDrinkOrder).then(dO => {
+//     setDataId(dO)
+//     setDrinkOrderIdDom(orderId, dO)
+//   })
+//   orderObj.drink_orders.push(newDrinkOrder);
+//   orderObj.drink_ids.push(drinkId)
+// }
+
+// function setDrinkOrderIdDom(order_id, drinkOrderObj) {
+//   Order.getOrderById(order_id).drink_orders.map(dO => {
+//     if (dO.id === undefined) {
+//       dO.id = drinkOrderObj.id
+//     }
+//   })
+// }
 
 function setDataId(obj) {
   document.querySelectorAll('[data-id=undefined]').forEach(element => element.dataset.id = obj.id)
